requirements.txt
azure-ai-ml==1.27.1
azure-identity==1.23.0
marimo==0.14.9
python-dotenv==1.1.1
